+++ langworld_db_data/removers/listed_value_remover.py
@@ -1,13 +1,12 @@
 from langworld_db_data import ObjectWithPaths
 from langworld_db_data.constants.literals import (
-    ID_SEPARATOR,
     KEY_FOR_FEATURE_ID,
     KEY_FOR_ID,
     KEY_FOR_VALUE_ID,
     KEY_FOR_VALUE_TYPE,
 )
 from langworld_db_data.tools.files.csv_xls import read_dicts_from_csv, write_csv
-from langworld_db_data.tools.value_ids.extract import extract_feature_id, extract_value_index
+from langworld_db_data.tools.ids.extract import extract_feature_id, extract_value_index
 
 
 class ListedValueRemoverError(Exception):
